db/requirements.txt
psycopg2==2.8.4
psycopg2-binary==2.8.6
google-cloud-logging==3.1.2